changed function structure
