+++ displayFunctions.js
@@ -1,0 +1,6 @@
+function resetBoard(rows, cols) {
+  const board = getBoard(rows, cols);
+  $('#grid').css('background-color', 'darkblue');
+
+  return board;
+}
