+++ index.html
@@ -39,7 +39,8 @@
 
     </div>
 
-    <script src="functions.js" type="text/javascript"></script>0
+    <script src="logicalFunctions.js" type="text/javascript"></script>
+    <script src="displayFunctions.js" type="text/javascript"></script>
     <script src="connect4.js" type="text/javascript"></script>
